temp commit / feat data storage overlay

+++ src/main/java/cn/howxu/mmcr/api/capability/CapabilityRequest.java
@@ -12,5 +12,5 @@ public interface CapabilityRequest {
 
     IOType ioType();
 
-    int parallelism();
+    long parallelism();
 }
